moved api for plugins to plugins module
