+++ mu4/plugins/api/cursor.cpp
