+++ mu4/plugins/api/selection.h
