+++ mu4/plugins/api/util.cpp
