+++ mu4/plugins/pluginsmodule.cpp
@@ -26,6 +26,7 @@
 #include "internal/pluginsservice.h"
 #include "internal/pluginsconfiguration.h"
 #include "view/pluginsmodel.h"
+#include "api/qmlpluginapi.h"
 
 using namespace mu::plugins;
 
@@ -55,6 +56,8 @@ void PluginsModule::registerResources()
 void PluginsModule::registerUiTypes()
 {
     qmlRegisterType<PluginsModel>("MuseScore.Plugins", 1, 0, "PluginsModel");
+
+    Ms::PluginAPI::PluginAPI::registerQmlTypes();
 
     framework::ioc()->resolve<framework::IUiEngine>(moduleName())->addSourceImportPath(plugins_QML_IMPORT);
 }
